+++ packages/blocks/core/js/third-party/blocksy-tax-query/test/tax-query-block.blocksy.e2e.cy.js
@@ -1,0 +1,71 @@
+/**
+ * Blockera dependencies
+ */
+import {
+	appendBlocks,
+	createPost,
+	openInnerBlocksExtension,
+	openMoreFeaturesControl,
+	savePage,
+	redirectToFrontPage,
+} from '@blockera/dev-cypress/js/helpers';
+
+describe('Blocksy → Advanced Taxonomies (Tax Query) Block → Block support', () => {
+	beforeEach(() => {
+		createPost();
+	});
+
+	it('Block support + Should not have inner blocks', () => {
+		appendBlocks(`<!-- wp:blocksy/tax-query {"uniqueId":"d27d7623"} -->
+<div class="wp-block-blocksy-tax-query"><!-- wp:blocksy/tax-template {"layout":{"type":"grid","columnCount":3}} -->
+<!-- wp:columns {"verticalAlignment":"center"} -->
+<div class="wp-block-columns are-vertically-aligned-center"><!-- wp:column {"verticalAlignment":"center","width":"33.33%"} -->
+<div class="wp-block-column is-vertically-aligned-center" style="flex-basis:33.33%"><!-- wp:blocksy/dynamic-data {"field":"wp:term_image","aspectRatio":"1","has_field_link":"yes"} /--></div>
+<!-- /wp:column -->
+
+<!-- wp:column {"verticalAlignment":"center","width":"66.66%"} -->
+<div class="wp-block-column is-vertically-aligned-center" style="flex-basis:66.66%"><!-- wp:blocksy/dynamic-data {"tagName":"h5","field":"wp:term_title","style":{"spacing":{"margin":{"bottom":"0px"}}},"has_field_link":"yes"} /-->
+
+<!-- wp:blocksy/dynamic-data {"field":"wp:term_count","after":" items"} /--></div>
+<!-- /wp:column --></div>
+<!-- /wp:columns -->
+<!-- /wp:blocksy/tax-template --></div>
+<!-- /wp:blocksy/tax-query -->
+		`);
+
+		cy.getBlock('blocksy/tax-query').click();
+
+		cy.getByAriaLabel('Select parent block: Advanced Taxonomies').click();
+
+		cy.get('.blockera-extension-block-card').should('be.visible');
+
+		cy.get('.blockera-extension.blockera-extension-inner-blocks').should(
+			'not.exist'
+		);
+
+		//
+		// 1. Edit Blocks
+		//
+
+		//
+		// 1.0. Block BG
+		//
+		cy.setColorControlValue('BG Color', '#dcffdc');
+
+		cy.getBlock('blocksy/tax-query').should(
+			'have.css',
+			'background-color',
+			'rgb(220, 255, 220)'
+		);
+
+		//
+		// 2. Assert styles in front end
+		//
+		savePage();
+		redirectToFrontPage();
+
+		cy.get('.blockera-block')
+			.first()
+			.should('have.css', 'background-color', 'rgb(220, 255, 220)');
+	});
+});